refactor introduction page
