+++ faverton-nuxt3/app/components/faverton/card/info/FavertonCardInfoPoints.vue
