+++ faverton-nuxt3/app/components/simulation/history/card/SimulationHistoryCard.vue
