+++ faverton-nuxt3/app/components/simulation/parameters/SimulationParameters.vue
