+++ faverton-nuxt3/app/components/simulation/result/monthly-amount/SimulationResultMonthlyAmount.vue
