+++ faverton-nuxt3/app/components/simulation/result/yearly-amount/SimulationResultYearlyAmount.vue
