--- faverton-nuxt3/app/composables/useSupabaseProfiles.ts
@@ -1,14 +1,0 @@
-export const useSupabaseProfiles = async () => {
-  const supabase = useSupabaseClient();
-  const user = useSupabaseUser();
-  const { data: profiles, error } = await supabase
-    .from(`profiles`)
-    .select(`*`).eq(`id`, user.value?.id || ``)
-    .single();
-
-  if (error) {
-    throw error;
-  }
-
-  return profiles;
-};
