--- faverton-nuxt3/app/pages/introduction/restaurants-cultural-activities/index.client.vue
@@ -1,17 +1,0 @@
-<script lang="ts" setup>
-const data = {
-  title: `Restaurants & Cultural Activities`,
-  subTitle: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.`,
-  button: `Learn more`,
-};
-</script>
-
-<template>
-  <div>
-    <IntroSourceEnergies
-      :title="data.title"
-      :sub-title="data.subTitle"
-      :button="data.button"
-    />
-  </div>
-</template>
