+++ faverton-nuxt3/app/pages/simulator/history/index.vue
@@ -1,6 +1,4 @@
 <script setup lang="ts">
-import CalcSimulationHistory from '~/components/calc/simulation/history/CalcSimulationHistory.vue';
-
 definePageMeta({
   middleware: [`auth`],
 });
@@ -8,6 +6,6 @@ definePageMeta({
 
 <template>
   <div>
-    <CalcSimulationHistory />
+    <SimulationHistory />
   </div>
 </template>
